update confirm delete , dashbroad success

+++ api_admin_laravel/resources/views/admin/page/feedback/feedback-list.blade.php
@@ -258,25 +258,42 @@
        
 $('body').on('click' ,'#deleteFeedback' , function(){
     var feedback_id = $(this).data("id");
-     if ( confirm("Bạn có chắc chắn muốn xóa phản hồi này không ?")) {
-    $.ajax({
-        type:"DELETE",
-        headers: {
-            'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
-        },
-        url:"{{ route('feedback.list.api') }}"+"/"+feedback_id,
-        success: function(){
-            table.ajax.reload();
-        },
-        error:function () {
-            console.log("xóa thất bại");
-        }
-    })
-     }
+    Swal.fire({
+                    title: "Bạn có chắc chắn?",
+                    text: "Bạn sẽ không thể hoàn tác!",
+                    icon: "warning",
+                    showCancelButton: !0,
+                    cancelButtonText: 'Quay lại',
+                    confirmButtonText: "Đúng, Xóa!",
+                    customClass: {
+                        confirmButton: "btn btn-primary",
+                        cancelButton: "btn btn-outline-danger ms-1",
+                    },
+                    buttonsStyling: !1,
+                    }).then(function (t) {
+                        if (t.value) {
+                            $.ajax({
+                                type:"DELETE",
+                                url:"{{ route('feedback.list.api') }}"+"/"+feedback_id,
+                                headers: {
+                                        'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
+                                    },
+                            success: function(){
+                                table.ajax.reload();
+                                toastr.success("Xóa Thành Công");
+                            },
+                            error:function () {
+                                toastr.error("Xóa không Thành Công");
+                            }
+                        })
+                        } 
+                    });
+    
 });
 
 $('body').on('click' ,'#editUser' , function(){
     var user_id = $(this).data("id");
+    
     $.ajax({
         type:"PATCH",
         url:"{{ route('user.list.api') }}"+"edit/"+user_id,
